+++ NEWS.md
@@ -13,6 +13,7 @@ Updated:
   * Replaced the package writexls with openxls for more functionality
   * Plotting: changed facet wrap labels, changed trending symbols
   * WSC station numbers with lower-case letters now accepted (08nm116 and 08NM116 both work)
+  * basin_area argument as list now will default to HYDAT area if not listed, otherwise NA
   
 #### Bugs Fixed: